Removed possibilty that points could generate on lattice edge
Now, when generating points with multiple lattices, if the point is too close to an adjoining lattice, it will not generate. This prevents, e.g., two points generating directly next to each other on opposite sides of a lattice boundary. "Too close" is a quarter of the smallest unit length between the two volumes.